src/Services/Agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "AI Agent service for mBot - handles AI interactions via RabbitMQ messaging"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faststream[rabbit]>=0.6.3",
    "pydantic-ai-slim[duckduckgo,google,groq,mcp,mistral,openai]>=1.14.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.4",
    "mypy>=1.13.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["contracts", "services"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
